requirements.txt
# Core Machine Learning & Data Processing
torch==2.1.0
tensorflow==2.14.0
transformers==4.34.0
pandas==2.1.1
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.2

# Financial Data & APIs
yfinance==0.2.32
polygon-api-client==1.14.0
ta==0.10.2
ccxt==4.0.36

# Pydantic & AI
pydantic==2.5.0
anthropic==0.7.0
openai==1.3.5

# HTTP & Async
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.0

# Optimization & Hyperparameter Tuning
optuna==3.14.0
ray==2.8.1
cvxpy==1.4.1

# Visualization & Reporting
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.18.0

# CLI & Configuration
rich==13.7.0
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0

# Database & Storage
sqlalchemy==2.0.23
redis==5.0.1
pymongo==4.6.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.3.0
python-json-logger==2.0.7

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Additional Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3
